pyproject.toml
[project]
name = "processor"
version = "0.1.0"
description = "CISON processor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.0",
    "pika>=1.3.2",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "sqlmodel>=0.0.31",
]
